+++ apps/auth-server/src/routes/auth.route/status.ts
@@ -1,0 +1,8 @@
+import { Router } from 'express';
+import { handleStatusCheck } from '@/controller/auth.controller/status';
+
+const statusAuthRouter = Router();
+
+statusAuthRouter.get('/', handleStatusCheck);
+
+export default statusAuthRouter;
